Fix default value for packet payload
The default value instantiation expression is not evaluated at runtime,
but only once. Thus, the default Payload object such as was suggested
would have been shared between all Packet instances instantiated with
this default parameter value, which is certainly not the semantic
intended.

The usage of None as a placeholder for instantiating a default value is
a common pattern.