Guard Socket.close() and Socket.port getter
The previous version unduly removed the closed status guard on the port
getter, because of exceptions raised by certain tests. Further analysis
of the results of these tests suggest that the problem arose when
finalizing socket instances that had been previously closed. This showed
that the problem was in forcing Socket.close() execution even on Socket
instances that had previously been closed. The guard has thus been
restored on the port getter, and another one has been set up on
Socket.close.